requirements.txt
pandas
tables
numpy
tqdm
einops
torchcde
scikit-learn
networkx
torch_scatter